// apps/site-import/src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigDecimalVP: { input: any; output: any; }
  BigInt: { input: any; output: any; }
  Text: { input: any; output: any; }
};

export type ExecutionStrategy = {
  __typename?: 'ExecutionStrategy';
  address: Scalars['String']['output'];
  axiom_snapshot_address?: Maybe<Scalars['String']['output']>;
  axiom_snapshot_slot?: Maybe<Scalars['BigInt']['output']>;
  destination_address?: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  quorum: Scalars['BigDecimalVP']['output'];
  timelock_delay: Scalars['BigInt']['output'];
  timelock_veto_guardian?: Maybe<Scalars['String']['output']>;
  treasury?: Maybe<Scalars['String']['output']>;
  treasury_chain?: Maybe<Scalars['Int']['output']>;
  type: Scalars['String']['output'];
};
// ...
export type Leaderboard = {
  __typename?: 'Leaderboard';
  id: Scalars['String']['output'];
  proposal_count: Scalars['Int']['output'];
  space: Space;
  user: User;
  vote_count: Scalars['Int']['output'];
};
// ...
export type Proposal = {
  __typename?: 'Proposal';
  author: User;
  axiom_snapshot_address?: Maybe<Scalars['String']['output']>;
  axiom_snapshot_slot?: Maybe<Scalars['BigInt']['output']>;
  cancelled: Scalars['Boolean']['output'];
  completed: Scalars['Boolean']['output'];
  created: Scalars['Int']['output'];
  edited?: Maybe<Scalars['Int']['output']>;
  executed: Scalars['Boolean']['output'];
  execution_destination?: Maybe<Scalars['String']['output']>;
  execution_hash: Scalars['String']['output'];
  execution_ready: Scalars['Boolean']['output'];
  execution_strategy: Scalars['String']['output'];
  execution_strategy_type: Scalars['String']['output'];
  execution_time: Scalars['Int']['output'];
  execution_tx?: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  max_end: Scalars['Int']['output'];
  metadata?: Maybe<ProposalMetadataItem>;
  min_end: Scalars['Int']['output'];
  proposal_id: Scalars['Int']['output'];
  quorum: Scalars['BigInt']['output'];
  scores_1: Scalars['BigDecimalVP']['output'];
  scores_2: Scalars['BigDecimalVP']['output'];
  scores_3: Scalars['BigDecimalVP']['output'];
  scores_total: Scalars['BigDecimalVP']['output'];
  snapshot: Scalars['Int']['output'];
  space: Space;
  start: Scalars['Int']['output'];
  strategies: Array<Maybe<Scalars['String']['output']>>;
  strategies_indices: Array<Maybe<Scalars['Int']['output']>>;
  strategies_indicies: Array<Maybe<Scalars['Int']['output']>>;
  strategies_params: Array<Maybe<Scalars['String']['output']>>;
  timelock_delay?: Maybe<Scalars['Int']['output']>;
  timelock_veto_guardian?: Maybe<Scalars['String']['output']>;
  tx: Scalars['String']['output'];
  type: Scalars['String']['output'];
  veto_tx?: Maybe<Scalars['String']['output']>;
  vetoed: Scalars['Boolean']['output'];
  vote_count: Scalars['Int']['output'];
};

export type ProposalMetadataItem = {
  __typename?: 'ProposalMetadataItem';
  body?: Maybe<Scalars['Text']['output']>;
  choices: Array<Maybe<Scalars['String']['output']>>;
  discussion?: Maybe<Scalars['Text']['output']>;
  execution?: Maybe<Scalars['Text']['output']>;
  id: Scalars['String']['output'];
  labels: Array<Maybe<Scalars['String']['output']>>;
  title?: Maybe<Scalars['Text']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  _checkpoint?: Maybe<_Checkpoint>;
  _checkpoints?: Maybe<Array<Maybe<_Checkpoint>>>;
  _metadata?: Maybe<_Metadata>;
  _metadatas?: Maybe<Array<Maybe<_Metadata>>>;
  executionstrategies?: Maybe<Array<Maybe<ExecutionStrategy>>>;
  executionstrategy?: Maybe<ExecutionStrategy>;
  leaderboard?: Maybe<Leaderboard>;
  leaderboards?: Maybe<Array<Maybe<Leaderboard>>>;
  proposal?: Maybe<Proposal>;
  proposalmetadataitem?: Maybe<ProposalMetadataItem>;
  proposalmetadataitems?: Maybe<Array<Maybe<ProposalMetadataItem>>>;
  proposals?: Maybe<Array<Maybe<Proposal>>>;
  space?: Maybe<Space>;
  spacemetadataitem?: Maybe<SpaceMetadataItem>;
  spacemetadataitems?: Maybe<Array<Maybe<SpaceMetadataItem>>>;
  spaces?: Maybe<Array<Maybe<Space>>>;
  strategiesparsedmetadatadataitem?: Maybe<StrategiesParsedMetadataDataItem>;
  strategiesparsedmetadatadataitems?: Maybe<Array<Maybe<StrategiesParsedMetadataDataItem>>>;
  strategiesparsedmetadataitem?: Maybe<StrategiesParsedMetadataItem>;
  strategiesparsedmetadataitems?: Maybe<Array<Maybe<StrategiesParsedMetadataItem>>>;
  user?: Maybe<User>;
  users?: Maybe<Array<Maybe<User>>>;
  vote?: Maybe<Vote>;
  votemetadataitem?: Maybe<VoteMetadataItem>;
  votemetadataitems?: Maybe<Array<Maybe<VoteMetadataItem>>>;
  votes?: Maybe<Array<Maybe<Vote>>>;
  votingpowervalidationstrategiesparsedmetadataitem?: Maybe<VotingPowerValidationStrategiesParsedMetadataItem>;
  votingpowervalidationstrategiesparsedmetadataitems?: Maybe<Array<Maybe<VotingPowerValidationStrategiesParsedMetadataItem>>>;
};
// ...
export type Space = {
  __typename?: 'Space';
  authenticators: Array<Maybe<Scalars['String']['output']>>;
  controller: Scalars['String']['output'];
  created: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  max_voting_period: Scalars['Int']['output'];
  metadata?: Maybe<SpaceMetadataItem>;
  min_voting_period: Scalars['Int']['output'];
  next_strategy_index: Scalars['Int']['output'];
  proposal_count: Scalars['Int']['output'];
  proposal_threshold: Scalars['BigDecimalVP']['output'];
  proposals: Array<Maybe<Proposal>>;
  proposer_count: Scalars['Int']['output'];
  strategies: Array<Maybe<Scalars['String']['output']>>;
  strategies_indices: Array<Maybe<Scalars['Int']['output']>>;
  strategies_indicies: Array<Maybe<Scalars['Int']['output']>>;
  strategies_metadata: Array<Maybe<Scalars['String']['output']>>;
  strategies_params: Array<Maybe<Scalars['String']['output']>>;
  strategies_parsed_metadata: Array<Maybe<StrategiesParsedMetadataItem>>;
  turbo: Scalars['Boolean']['output'];
  tx: Scalars['String']['output'];
  validation_strategy: Scalars['String']['output'];
  validation_strategy_params: Scalars['Text']['output'];
  verified: Scalars['Boolean']['output'];
  vote_count: Scalars['Int']['output'];
  voter_count: Scalars['Int']['output'];
  voting_delay: Scalars['Int']['output'];
  voting_power_validation_strategies_parsed_metadata: Array<Maybe<VotingPowerValidationStrategiesParsedMetadataItem>>;
  voting_power_validation_strategy_metadata: Scalars['String']['output'];
  voting_power_validation_strategy_strategies: Array<Maybe<Scalars['String']['output']>>;
  voting_power_validation_strategy_strategies_params: Array<Maybe<Scalars['String']['output']>>;
};

export type SpaceMetadataItem = {
  __typename?: 'SpaceMetadataItem';
  about: Scalars['String']['output'];
  avatar: Scalars['String']['output'];
  cover: Scalars['String']['output'];
  delegations: Array<Maybe<Scalars['String']['output']>>;
  discord: Scalars['String']['output'];
  executors: Array<Maybe<Scalars['String']['output']>>;
  executors_destinations: Array<Maybe<Scalars['String']['output']>>;
  executors_strategies: Array<Maybe<ExecutionStrategy>>;
  executors_types: Array<Maybe<Scalars['String']['output']>>;
  external_url: Scalars['String']['output'];
  github: Scalars['String']['output'];
  id: Scalars['String']['output'];
  labels: Array<Maybe<Scalars['String']['output']>>;
  name: Scalars['String']['output'];
  treasuries: Array<Maybe<Scalars['String']['output']>>;
  twitter: Scalars['String']['output'];
  voting_power_symbol: Scalars['String']['output'];
  wallet: Scalars['String']['output'];
};
// ...
export type StrategiesParsedMetadataDataItem = {
  __typename?: 'StrategiesParsedMetadataDataItem';
  decimals: Scalars['Int']['output'];
  description: Scalars['String']['output'];
  id: Scalars['String']['output'];
  name: Scalars['String']['output'];
  payload?: Maybe<Scalars['String']['output']>;
  symbol: Scalars['String']['output'];
  token?: Maybe<Scalars['String']['output']>;
};
// ...
export type StrategiesParsedMetadataItem = {
  __typename?: 'StrategiesParsedMetadataItem';
  data?: Maybe<StrategiesParsedMetadataDataItem>;
  id: Scalars['String']['output'];
  index: Scalars['Int']['output'];
  space: Space;
};
// ...
export type User = {
  __typename?: 'User';
  address_type: Scalars['Int']['output'];
  created: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  proposal_count: Scalars['Int']['output'];
  proposals: Array<Maybe<Proposal>>;
  vote_count: Scalars['Int']['output'];
  votes: Array<Maybe<Vote>>;
};
// ...
export type Vote = {
  __typename?: 'Vote';
  choice: Scalars['Int']['output'];
  created: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  metadata?: Maybe<VoteMetadataItem>;
  proposal: Scalars['Int']['output'];
  space: Space;
  tx: Scalars['String']['output'];
  voter: User;
  vp: Scalars['BigDecimalVP']['output'];
};

export type VoteMetadataItem = {
  __typename?: 'VoteMetadataItem';
  id: Scalars['String']['output'];
  reason: Scalars['Text']['output'];
};
// ...
export type VotingPowerValidationStrategiesParsedMetadataItem = {
  __typename?: 'VotingPowerValidationStrategiesParsedMetadataItem';
  data?: Maybe<StrategiesParsedMetadataDataItem>;
  id: Scalars['String']['output'];
  index: Scalars['Int']['output'];
  space: Space;
};
// ...
export type _Checkpoint = {
  __typename?: '_Checkpoint';
  block_number: Scalars['Int']['output'];
  contract_address: Scalars['String']['output'];
  /** id computed as last 5 bytes of sha256(contract+block) */
  id: Scalars['ID']['output'];
};
// ...
export type _Metadata = {
  __typename?: '_Metadata';
  /** example: last_indexed_block */
  id: Scalars['ID']['output'];
  value?: Maybe<Scalars['String']['output']>;
};
// ...
export type ProposalQuery = { __typename?: 'Query', proposal?: { __typename?: 'Proposal', id: string, proposal_id: number, quorum: any, execution_hash: string, start: number, min_end: number, max_end: number, snapshot: number, scores_1: any, scores_2: any, scores_3: any, scores_total: any, execution_time: number, execution_strategy: string, execution_strategy_type: string, execution_destination?: string | null, timelock_veto_guardian?: string | null, strategies_indices: Array<number | null>, strategies: Array<string | null>, strategies_params: Array<string | null>, created: number, edited?: number | null, tx: string, execution_tx?: string | null, veto_tx?: string | null, vote_count: number, execution_ready: boolean, executed: boolean, vetoed: boolean, completed: boolean, cancelled: boolean, space: { __typename?: 'Space', id: string, controller: string, authenticators: Array<string | null>, metadata?: { __typename?: 'SpaceMetadataItem', id: string, name: string, avatar: string, voting_power_symbol: string, treasuries: Array<string | null>, executors: Array<string | null>, executors_types: Array<string | null>, executors_strategies: Array<{ __typename?: 'ExecutionStrategy', id: string, address: string, destination_address?: string | null, type: string, treasury_chain?: number | null, treasury?: string | null } | null> } | null, strategies_parsed_metadata: Array<{ __typename?: 'StrategiesParsedMetadataItem', index: number, data?: { __typename?: 'StrategiesParsedMetadataDataItem', id: string, name: string, description: string, decimals: number, symbol: string, token?: string | null, payload?: string | null } | null } | null> }, author: { __typename?: 'User', id: string, address_type: number }, metadata?: { __typename?: 'ProposalMetadataItem', id: string, title?: any | null, body?: any | null, discussion?: any | null, execution?: any | null, choices: Array<string | null>, labels: Array<string | null> } | null } | null };